Refactor src into logical packages and implement core logic enhancements.
- Reorganized src/ into models, events, core, ui, cli, and utils.
- Updated all include paths and test suites (Unit, BDD, SDD).
- Implemented core logic for task metadata, archiving, and actual effort tracking.
- Ensured all existing tests pass within the new architecture.
- Removed binary artifacts from the repository.
