+++ queue/pending/task1.json
@@ -1,0 +1,1 @@
+{"task_id": "task1","description": "desc","priority": "high","dependencies": [],"component": "comp","max_runtime_sec": 10,"date": "","time": "","platform": "","service": "","confirmed": false,"overlapping_task_ids": [],"first_name": "","last_name": "","contact_info": "","anonymous_id": "","labels": [],"due_date": "","estimated_effort": 0,"actual_effort": 0,"blocked_by_note": "","owner": "","watchers": [],"archived": false,"sequence_number": 0,"cancellation_reason": "","creation_time": ""}
