--- src/pubsub.h
@@ -1,29 +1,0 @@
-#ifndef PUBSUB_H
-#define PUBSUB_H
-
-#include <vector>
-#include <algorithm>
-#include <map>
-
-enum class EventType;
-struct Event;
-
-// Subscriber interface
-class ISubscriber {
-public:
-    virtual ~ISubscriber() = default;
-    virtual void onEvent(const Event& event) = 0;
-};
-
-// Publisher class
-class Publisher {
-public:
-    void subscribe(EventType eventType, ISubscriber* subscriber);
-    void unsubscribe(EventType eventType, ISubscriber* subscriber);
-    void publish(const Event& event);
-
-private:
-    std::map<EventType, std::vector<ISubscriber*>> subscribers;
-};
-
-#endif // PUBSUB_H
